+++ libsolidity/codegen/ir/IRGeneratorForStatements.cpp
@@ -1090,12 +1090,11 @@ void IRGeneratorForStatements::endVisit(FunctionCall const& _functionCall)
 						")\n";
 				else if (auto functionType = dynamic_cast<FunctionType const*>(paramTypes[i]))
 				{
-					solAssert(
-						IRVariable(arg).type() == *functionType &&
-						functionType->kind() == FunctionType::Kind::External &&
-						!functionType->hasBoundFirstArgument(),
-						""
-					);
+					auto const* argFunctionType = dynamic_cast<FunctionType const*>(&IRVariable(arg).type());
+					solAssert(argFunctionType);
+					solAssert(argFunctionType->kind() == FunctionType::Kind::External);
+					solAssert(!argFunctionType->hasBoundFirstArgument());
+					solAssert(functionType->kind() == FunctionType::Kind::External);
 					define(indexedArgs.emplace_back(m_context.newYulVariable(), *TypeProvider::fixedBytes(32))) <<
 						m_utils.combineExternalFunctionIdFunction() <<
 						"(" <<